Add files via upload
Dark1 is still working. Improved the logic for positioning blocks in the board: now you can't put opaque or reflective blocks next to a target (since it would prevent the laser from hitting it)... now that I'm thinking about it I should also account for the direction of the laser, because if the laser is coming from the left then you could put a block to the right of the target, but not to the left. Same for top and bottom.