Add files via upload
Fixed the infinite while loo in solve_game when placing blocks by re-defining next_rand inside the loop. Added the iterations count which is not fully working but it should at least put a cap on the running time for the program (while loop will run until the game is solved or until the max iteration number has been reached). Still have to fix the block placement reasoning for blocks surrounding targets.